server/requirements.txt
asgiref==3.7.2
Django==4.2.8
django-six==1.0.5
djangorestframework==3.14.0
mysqlclient==2.2.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.3
xlwt==1.3.0